+++ packages/react/src/themes/teams-dark/componentVariables.ts
@@ -17,5 +17,5 @@ export { default as Reaction } from './components/Reaction/reactionVariables'
 export { default as Alert } from './components/Alert/alertVariables'
 export { default as ProviderBox } from './components/Provider/providerBoxVariables'
 export { default as Dropdown } from './components/Dropdown/dropdownVariables'
-
+export { default as Label } from './components/Label/labelVariables'
 export { default as PopupContent } from './components/Popup/popupContentVariables'